refactor(router): move in-app deep link handling out of build()
The handler for deep links delivered while the app is running was a
414-line closure inside _DivineAppState.build() — a switch over every
DeepLinkType, with the navigation for each inlined. Nothing could reach it
without building the whole app, so the only parts under test were the two
decision helpers that had already been lifted to top-level functions.

Move it to DeepLinkCoordinator, along with those two helpers and their
enums, which belong with it rather than in the app entrypoint. The closure
captured nothing but `ref` — twice — so the coordinator takes a GoRouter and
an AuthService and the extraction is otherwise verbatim.

DeepLinkCoordinator carries a note saying why the duplication with
appRouterRedirect must stay: Android can hand a custom scheme to GoRouter
before the app-links stream sees it, both paths are deliberately idempotent,
and returning early from the redirect previously skipped the minor-account
and auth gates (#5195, #7146). Verified on device — divine://inbox reaches
both, and only the router resolves it.

Their two test files move from test/main_* to test/router/, where they now
mirror the source they cover, and a new deep_link_coordinator_test asserts
where each link type actually lands.

main.dart: 3450 -> 2782 lines.
